Docker: don't break existing images on tag change
Previously, DockerClient.GetInstance would return an error if the instance's
image does not match the supplied value. This can happen naturally if, for
example, the major.minor tag for an image now points to a newer
major.minor.patch than it did when the Dockerized instance was originally
created.

This commit improves the logic to fall back on checking the instance's Flavor,
and only erroring if that too does not match the expectation.