workspace=docker: Improve image verification logic
When using workspace=docker, if Skeema finds a pre-existing container created
from a previous run, the image of the container is verified. The previous
logic was overly strict, erroring in two valid situations:

* If the local Docker tag for the major.minor version now points to a
  different patch release, due to other non-Skeema Docker usage on the system

* If the version of Docker is several years old and uses the previous
  hashing scheme for image identification

In either of these cases, Skeema no longer errors, as long as the Dockerized
instance's flavor (vendor and major.minor version) can be introspected and
correctly match that of the desired image (via Skeema's flavor option).