Routines parser: don't fail on empty single-line comments
Due to a subtle regular expression mistake, previously the parser would
erroneously fail empty comments within the body of a proc or func, e.g.
tokens of "#\n" or "-- \n".

This situation was somewhat rare because many MySQL clients (including the
official one) perform client-side stripping of comments from statements by
default.

With any luck this will be a solution to #96.