Adding --ddl-wrapper-comment for diff output A new flag is introduced, --ddl-wrapper-comment (bool, default `false` for backward compatibility), which controls output of `skeema diff` (or `skeema push --dry-run`) command.
When this flag is given, each DDL is wrapped by well formed comments, in the likeness of:
```
-- skeema:ddl:begin
CREATE TABLE `table_0` (
  `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
  `name` varchar(128) NOT NULL,
  PRIMARY KEY (`id`)
) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
-- skeema:ddl:end
```

The purpose of this flag is to assist with parsing of `skeema diff` output. As a concrete example, we want to be able to run `skeema diff` and review the statements generated by the tool, on a per-statement basis. When `diff` contains multiple DDL statements (multiple `CREATE`, `ALTER`, `DROP`) it is difficult task to parse it. Delimiting by semicolon (`;`) is unsafe, because a semicolon can appear as, say, a `VARCHAR`s default value, or in a column comment.

With the above wrapper comments the parsing operation is safe and simple. The wrapper comments text is extremely unlikely to appear as a valid text in a DDL.