RunSuite: Speed up filtered integration test runs
Golang supports `go test -run ...` to filter test execution, and this can
be used to filter subtests as well. For example `go test -run Integ/Instance`
will only run integration tests containing the word Instance in the test
method name.

Previously this was rather slow, because the integration test suite's
BeforeTest() method was still executed for every skipped test. This commit
adds logic to prevent unnecessary redundant BeforeTest() calls.

This commit also changes the signature of BeforeTest() to remove the method
name arg, which is a breaking change, but necessary since the method name arg
could otherwise now be incorrect. This package is still pre-1.0 in terms of
SemVer.