Add initial diff logic for partitioning status
This commit adds two new structs satisfying the TableAlterClause interface:
AddPartitioning and RemovePartitioning, returned as part of a TableDiff when
a change in partitioning status is requested.

A new StatementModifiers field, Partitioning, is added (along with
corresponding enum constant values) to control whether or not partitioning
status changes are suppressed in different situations. This also permits
suppressing partitioning clauses when creating new tables as well.