diff/push: Ignore LF vs CRLF differences in routines' param lists
When diffing procs and funcs, the intended behavior is to treat UNIX-style
line endings (LF) and Windows-style (CRLF) as equivalent. However, the
previous fix for this in commit 38f40f8bf did not convert line-endings inside
param lists in situations where mysql.proc is available. This could cause
confusing phantom diffs where the only change to a routine is in its line-
endings.

This bug did not affect MySQL 8 or Percona Server 8, since mysql.proc does not
exist in these versions and a different code path is used.