SchemaDiff: Include ALTERs to drop partitions before dropping table
When a diff includes a DROP TABLE for a table partitioned with method RANGE,
RANGE COLUMNS, LIST, or LIST COLUMNS, the DROP TABLE will now be preceded by
a series of ALTER TABLE ... DROP PARTITION statements to drop all but one of
the table's partitions prior to dropping the table.

This is done to avoid having a single long operation that holds MySQL's
dict_sys mutex, which can be detrimental to application workloads. (Same
reasoning as in Instance.DropTablesInSchema from commit 68179239b, but
different code path.)

This logic isn't implemented for other partitioning methods because they
do not support ALTER TABLE ... DROP PARTITION.