Initial partitioned table support
This commit updates dep github.com/skeema/tengo to point at its partitioning
branch, in order to bring in low-level support for introspection and diff of
partitioned tables. It also adds a new partitioning option to diff/push to
control behavior when interacting with partitioned tables.

Documentation will be added in a subsequent commit.