Deprecate option --reuse-temp-schema, and fix bugs in it
This commit deprecates the reuse-temp-schema option, since it is not commonly
used, and doesn't provide any real advantages over the default behavior of
dropping and recreating the temp schema each time.

This commit also fixes two edge case bugs in reuse-temp-schema's behavior:

* When used in an environment with multiple schemas per instance and
  differing schema-level default charset or collation values, a previously-
  processed schema's defaults could remain in-place on the temp schema,
  inadvertently affecting the introspected workspace for subsequent schemas.

* When used in the presence of stored procs/funcs, the procs/funcs would not
  be cleaned up between runs, typically causing the subsequent run to fail
  with duplicate routine errors.

Integration testing coverage has been added to ensure that the temp schema is
truly empty of all supported object types prior to use. This will avoid an
undetected regression if e.g. support for views or triggers is added to a
future version of Skeema.