Add option temp-schema-threads for configurable concurrency
This commit adds a new option, temp-schema-threads, which controls the
concurrency level for workspace CREATE and DROP queries when using
workspace=temp-schema. The default value for this option is 5 threads, which
is a reduction from the previous hardcoded default concurrency of 10 threads.

Users may wish to raise this value to improve Skeema's performance in some
scenarios, or lower this value to prevent mutex contention in other scenarios.
Refer to the option reference for more information about tuning this option.

This relates to #93. Keeping that issue open for now though since an
additional related improvement is also planned (ability to disable binlogging
for temp-schema operations)