Add new option temp-schema-binlog controlling workspace DDL replication
Previously, with workspace=temp-schema (the default), all workspace operations
were written to the binary log, meaning that replicas would execute them. This
can be undesirable for performance or operational cleanliness reasons, since
running the operations on replicas serves no real purpose: the temp schema is
dropped after each run, and Skeema won't interact with the replicas anyway.

This commit adds a new enum option, temp-schema-binlog, which can be set to
either "on" (keep the old behavior of writing these queries to binlog), "off"
(skip binlogging of workspace queries -- error if non-super-user), or the
default of "auto" (skip binlogging of workspace queries only if super-user).

This relates to issue #93.