schema option: permit regex values wrapped in slashes
This commit adds support for a new value type in the schema option: if the
value is wrapped in forward slashes, it is interpreted as a regular
expression, matched against the existing schema names on the target instance.

For example, this value in a .skeema file:
    schema=/^foo/
will map the dir to all existing schema names that begin with prefix "foo".

Previously this could be accomplished by combining schema=* with an
ignore-schema regex in the same dir, but that can be awkward in situations
where the special unsharded schemas are numerous and don't follow a common
naming pattern.