Add diff support for generated columns (MySQL 5.7+, MariaDB 10.2+)
This commit adds diff support for tables containing one or more generated
columns, whether virtual or stored.

A new StatementModifier field permits insertion of WITH VALIDATION clauses
to ALTER TABLE statements adding or modifying a virtual generated column.

Note that generated columns in MariaDB 10.1 are not supported, as that
release used different syntax for generated columns.