diff/push: support altering tables with generated columns
This commit adds diff support for tables using generated columns (virtual or
stored) in MySQL 5.7+, Percona Server 5.7+, and MariaDB 10.2+.

This commit also adds a new boolean option, alter-validate-virtual, which can
optionally be used to add a WITH VALIDATION clause to ALTERs affecting virtual
columns.

Resolves #49.

Note that although MariaDB 10.1 has support for generated columns, it uses a
different syntax and information_schema representation for them, which is not
supported by Skeema at this time.