Add new utility function ParseCreatePartitioning()
This commit adds a new regex-based method, ParseCreatePartitioning(), to split
the base of a CREATE TABLE statement from its partitioning clause.

The existing Table.UnpartitionedCreateStatement() method now uses the new
utility method in situations where the partitioning clause does not match the
canonical expectation or the accurate Flavor is not available.