Index visibility alteration: fix compat with StrictIndexOrder
Previous commit 515748af added support for MySQL 8.0 invisible indexes, and
generation of ALTER INDEX ... {VISIBLE|INVISIBLE} in diffs to change the
visibility of an existing index. However, in combination with a true value
for StatementModifiers.StrictIndexOrder, illegal DDL could be generated: an
ALTER TABLE which uselessly does an ALTER INDEX prior to dropping and re-
adding the same index for ordering purposes. This syntax is actually invalid,
as ALTER TABLE does not permit an ALTER INDEX clause if a DROP INDEX is
present for the same index, even if an ADD INDEX is also present.

This commit fixes the situation to suppress the ALTER INDEX in this specific
scenario.