MySQL 8.0.19: edge case fix for MODIFY COLUMN
This commit provides one additional fix relating to MySQL 8.0.19's removal
of display widths from integer types:

If an integer column only differs by presence or lack of display width,
suppress generation of a MODIFY COLUMN clause. Otherwise, when comparing a
pre-8.0.19 version of a table to a post-8.0.19 version, an extraneous MODIFY
COLUMN clause would be emitted.

This situation is rare, but can occur when using workspace=docker with a
different patch version than the production databases are using; and/or in a
sharded environment in which shards are running different patch versions.