diff/push: allow schema names in *.sql in some cases
Previously, inclusion of a schema name in a .sql file (such as a USE command,
or a dbname prefix in a CREATE statement) caused skeema diff and skeema push
to have undefined/broken behavior.

This commit allows directories with these SQL statements to function correctly
in a couple specific scenarios:

a) No schema name is defined in .skeema, and all CREATEs in the directory are
   preceded by a USE command or have a specific dbname prefix

or

b) A single schema name is defined in .skeema (or a single schema name is
   returned by a shellout or regex), and all CREATEs in the directory are
   preceded by a USE command for this same schema name, or use a dbname
   prefix matching this same schema name

This work is a stepping-stone towards eventually providing compatibility with
other logical dump tools, such as mysqldump or sqldef.