Instance.CanSkipBinlog: Fix bug with ALL PRIVILEGES
Whenever a user had ALL PRIVILEGES on at least one database or table, but not
the global version of ALL PRIVILEGES (ALL PRIVILEGES ON *.*), previously
Instance.CanSkipBinlog would erroneously return true. This commit fixes the
method's logic to properly handle this case, and adds additional test
coverage for this method.