temp-schema: fix errors with certain non-SUPER users
When running Skeema with a non-root user, and using default settings
workspace=temp-schema and temp-schema-binlog=auto, certain combinations of
privileges would erroneously cause fatal error "Cannot create temporary
schema on ...: Error 1227: Access denied; you need (at least one of) the
SUPER privilege(s) for this operation".

Specifically, this would occur if the user had ALL PRIVILEGES on at least
one database or table, but not the global version of ALL PRIVILEGES (ALL
PRIVILEGES ON *.*).

This commit updates to the latest version of the tengo dep, which includes a
fix for the bug's root cause.