lint, diff/push: Add new linter rule --check-compression
This commit adds support for a new linter rule that checks the compression
settings of tables, ensuring that each setting is found in companion list
option --allow-compression. This option can be used to restrict which types/
sizes of compression are permitted, or require all tables to be compressed,
or require that no tables are compressed.

Thank you to Etsy for generously sponsoring the development of this feature!