lint, format, pull: Clearer error message if no host defined
Prior to Skeema v1.4.0, `skeema lint` and `skeema format` provided an error
message when invoked on a directory that does not define a host, or when
invoked using an environment name that does not define a host. However, when
the temp-schema-binlog option was introduced in v1.4.0, this behavior
inadvertently changed to a nil pointer dereference panic.

This commit fixes detection of this situation for `skeema lint` and `skeema
format` to avoid the panic, as well as providing a clear message in `skeema
pull` (which previously just displayed no output in this case). Test coverage
is now included to prevent a future regression, and the workspace subpackage
also now contains a nil pointer check guard.

Fixes #141. Thank you to @razor-1 for the bug report!