Internal changes for OS portability
This commit includes several changes related to Windows compatibility. This
also breaks the ability to compile Skeema with GOOS=windows for now. However,
this is preferable to the previous situation, in which Skeema compiled
successfully but contained numerous portability bugs which prevented correct
or safe behavior on Windows.

A functional Windows port has now been completed, but several source code
files are being omitted here from the open source release. The Windows port
required a nontrivial investment (monetary as well as time/effort), and will
initially only be available for the commercial edition of Skeema, planned for
Q2 of 2021.