Index.RedundantTo: account for uniqueness constraints
Previously, Index.RedundantTo() would return true for a unique index
receiver in situations where another unique index had the same initial
columns plus other additional parts/columns. However, this logic did not
account for the fact that a unique index also represents a uniqueness
constraint, which has a different effect if the column list differs in any
way.

This commit fixes this logic bug, so that a unique index is now only
considered redundant to another unique index (or PK) if the parts list is
exactly identical between the two.