lint-dupe-index: respect uniqueness constraints
Previously, the linter's duplicate index checker would flag unique indexes
whose initial columns/parts are a left-to-right subset of another larger
unique index. However, this logic did not account for the fact that a unique
index also represents a uniqueness constraint, which may intentionally exist
as a subset of another unique index or the primary key.

This commit updates the tengo package to bring in the fix for this bug, so
that the dupe-index checker now only flags a unique index if it is an exact
duplicate of another unique index or the primary key.

Fixes #152. Thank you to @jearle-dealnews for the bug report!