pull, add-environment: don't mangle bools when rewriting .skeema
Whenever an existing .skeema option file is being rewritten during execution
of `skeema pull` (e.g. due to change in flavor) or `skeema add-environment`,
previously any boolean option values in the file would be mangled in a
confusing way. The options still worked properly but were hard to read;
enabled bool options would be set to a value of 1, and disabled bool options
would be set to an empty string.

This commit updates the mybase dep to fix this problem. Now bool options are
written in their canonical format: "foo" to enable, "skip-foo" to disable, no
equals sign or value present in either case.