diff/push: fix for MySQL 8 column default expressions
MySQL 8 supports using expressions for column defaults. Skeema correctly
handled some forms of expressions, but in other cases marked tables as
unsupported-for-diff. This commit updates the tengo dep to bring in a fix
for these cases, so that they are correctly supported for diffing.

This fix also correctly handles blob/text default expressions, even though
information_schema omits them due to a bug in MySQL 8.0.13-8.0.22.