Table diffs: improve performance in common case
Changing the position (order) of table columns is relatively rare, but
previously consumed a large portion of the table diff computation time. This
commit changes the logic to detect lack-of-reordering, and greatly simplify
the comparison logic if no columns were reordered.

For tables with high numbers of columns, benchmarking shows a roughly 75%
reduction in column comparison time with this change if no columns were
reordered.