MySQL 8.0.19+: fix lingering I_S int display width
Integer column type display widths were removed in MySQL 8.0.19 and no longer
appear in SHOW CREATE TABLE. This package's handling for this situation was
largely fixed one year ago in 97e9386 and a866906, however one edge case
remained: if a database is upgraded from a pre-8.0.19 version of MySQL 8, to
version 8.0.19+, information_schema still retains display widths until the
table is next altered. This does not occur if the DB was upgraded from pre-
8.0 directly to 8.0.19+, nor does it occur with tables initially created in
8.0.19+.

Previously, tables affected by this bug would be erroneously marked as
unsupported for diff. This commit handles the situation correctly, and also
slightly improves performance of the old logic by not using regexes.

This commit also includes an unrelated unit test that was accidentally
omitted from the previous commit.