diff/push: Fix lingering int display widths in MySQL 8.0.19+
Integer column type display widths were removed in MySQL 8.0.19 and no longer
appear in SHOW CREATE TABLE. Skeema's handling for this situation was largely
fixed one year ago in 533d073 and c7ee397. However, a bug with one specific
edge case apparently remained: if a database is upgraded from a pre-8.0.19
version of MySQL 8, to version 8.0.19+, information_schema still retains
display widths until the table is next altered. This confuses one of Skeema's
core safety checks, causing the table to be marked as unsupported for diff.
This commit fixes this bug.

This edge case does not occur if the DB was upgraded from pre-8.0 (such as
5.6 or 5.7) directly to 8.0.19+, nor does it occur with tables initially
created in 8.0.19+. Skeema's integration test suite runs against the newest
patch release for each supported flavor, but does not test the MySQL upgrade
process, so the exact situation cannot be reproduced in the test suite.
However, related logic is now covered by unit tests in Tengo, and the fix has
been manually confirmed in an upgrade scenario (from 8.0.18, to 8.0.23).

Fixes #158. Thank you to @trevorr for the issue report!