Instance.CanConnect: close pools immediately
Previous commit ad6d55f included a refactor to Instance.CanConnect to use the
new non-cached Instance.ConnectionPool method. When running heavy tests (such
as Skeema's integration test suite), the number of open connections steadily
grows due to the leaking pools here. This commit changes CanConnect to ensure
that if the cache was bypassed, the pool is closed immediately.