Improve connection management and performance
This commit includes several small changes relating to connection-handling
across all Skeema commands.

Skeema now auto-detects when the database has been configured to use a
relatively low value for max_user_connections or max_connections. Skeema's
query concurrency is throttled down accordingly when this is the case.
Extremely low limits (e.g. 10 or less) may still be problematic for now
though.

The number of network round-trips per new connection has been reduced. This
should provide a performance benefit for users with nontrivial network
latency between Skeema and the database.

This commit also includes minor changes relating to the flavor option in rare
situations where the flavor cannot be auto-detected from an instance. In this
case, if the flavor is supplied on the command-line to `skeema init`, it will
be persisted into the host-level .skeema file. If the configured flavor
includes a patch version number, it will now be used, whereas previously the
patch number was inadvertently ignored.