Fix failing test in applier subpackage
The logic in applier's TestTargetsForDirMulti previously relied on stopping
containerized test instances in order to simulate offline DBs. However, the
prior commit 00510a8 changed the instance validation logic to avoid re-testing
connectivity on a DB already known to recently be online/reachable, and this
breaks the assumptions in this test.

The test logic now uses incorrect port numbers to simulate offline instances,
as this is much simpler and faster to execute.