Schema introspection: parallelize I_S queries
To introspect a schema, Instance.Schemas() runs a number of queries on
information_schema. Previously these queries were run sequentially, which can
be slow over a WAN or VPN. This commit changes these queries so that most of
them run concurrently when possible (e.g. depending on max_user_connections).