Parallelize schema introspection queries
This commit updates the vendored copy of Tengo to bring in a commit which
parallelizes queries against information_schema during schema introspection.
Previously these queries were run sequentially, which can be slow over a WAN
or VPN.

This change may increase test suite execution time slightly, as well as
Skeema execution time when running against a database on localhost more
generally. The extra concurrency isn't beneficial when network latency is
nonexistent. This logic may be tweaked in the future to provide a better
balance.