Routines parser: fix lexer for --\n comments
Routines containing a blank comment of the form --\n caused the statement
lexer to incorrectly treat the following line as part of the comment. This
could then cause a parser failure if that following line contained the start
of a multi-line string.

Previous commit 94f380a in Aug 2019 handled related cases "#\n" and "-- \n",
but did not properly handle "--\n".

See #159. Thank you to @varadz for the bug report!