init, pull: record Skeema version in new generator option
This commit modifies `skeema init` and `skeema pull` to record the version
and edition of Skeema used to generate or update the directory structure.

This value will be used in the future to ensure backwards compatibility when
a user upgrades to a new version of Skeema, or switches between the FOSS
community edition and the commercial edition. This will allow diff/push to
determine whether a given type of objects (e.g. triggers, views, etc) is
missing from *.sql simply due to generation from an older Skeema that did not
support those objects, vs being missing because the user intentionally removed
those CREATEs in order to drop the objects.