Instance.DropTablesInSchema: fix two subtle bugs
1) If any DROP TABLE statements from DropTablesInSchema returned an error
other than deadlock, the error was inadvertently ignored previously due to
variable shadowing. The error now propagates properly.

2) MySQL 8 includes views in information_schema.partitions, which is a change
relative to earlier versions of MySQL (and all versions of MariaDB). With the
above fix, this then caused Instance.DropTablesInSchema() to return an error
if any views are present in the schema, due to assumptions made in the logic
of instance.go's tablesToPartitions().