BulkDropOptions: permit a *Schema to be supplied
This commit adds new field BulkDropOptions.Schema, optionally permitting a
pointer to a Schema to be supplied as part of the options. If present, this
may reduce the number of queries executed by the Instance drop methods. For
example, DropTablesInSchema will use the supplied schema's table list instead
of re-querying information_schema; ditto for DropRoutinesInSchema.

This field should only be used in cases where the schema was introspected very
recently and is known to still be accurate, in terms of existence of tables,
routines, etc.