workspace=docker: handle rare case of missing port mapping
When using workspace=docker, and either a new container is created or existing
stopped container is started, in some rare cases sometimes Skeema inspects
the container prior to its port mapping being ready. This could manifest as
ambiguous errors with connecting to the container. The problem has been
observed with MySQL 8.0.24 images on Docker for Mac; it is unclear if it
occurs in other situations.

This commit solves the problem by adding retries to the container inspect call
if the port number is not yet available. If the port number is still
unavailable after 5 retries, a clearer fatal error is now returned.