MySQL 8.0.24: fix handling of 3-byte utf8 charset
Last week's release of MySQL 8.0.24 included a change related to the legacy
3-byte utf8 character set: SHOW CREATE TABLE now outputs this character set
as "utf8mb3", but information_schema still returns "utf8". That change broke
this package's diff safety check for tables using that character set in
MySQL 8.0.24.

This commit handles the situation properly by using "utf8mb3" in place of
"utf8" if the database server is 8.0.24+.