MySQL 8.0.24: diff support for utf8mb3 tables
Last week's release of MySQL 8.0.24 included a change related to the legacy
3-byte utf8 character set: SHOW CREATE TABLE now outputs this character set
as "utf8mb3" at the level of a table's default charset. However, "utf8" is
still used in all other locations: columns in SHOW CREATE TABLE, and all
places in information_schema.

That MySQL 8.0.24 change broke a safety check used in `skeema diff` and
`skeema push`, causing tables with default charset of utf8mb3 to show up as
unsupported for diff. This commit fixes diff support for these tables.