TableDiff.alterStatement: non-InnoDB StrictIndexOrder fix
Due to a logic bug, when diffing non-InnoDB tables, the StrictIndexOrder
statement modifier would inadvertently be enabled for any table with a PK.
This occurred because Table.ClusteredIndexKey() always returns nil for non-
InnoDB tables.

This commit fixes the bug by adding an InnoDB engine check on the conditional.
The intention of this code is to only force StrictIndexOrder during situations
in which the ordering of non-nullable unique indexes is relevant, since InnoDB
will use the first one as the clustered index key if there's no explicit PK.
(Other storage engines typically may not share this behavior.)