lint-pk: explain problem with nullable UNIQUE KEY
The lint-pk checker flags any table that lacks an *explicit* PRIMARY KEY, even
if it has an implicit one. That behavior remains as-is, but this commit
improves the explanatory text for InnoDB tables with UNIQUE KEYs that *cannot*
be used as an *implicit* PK.

When an InnoDB table has no explicit PK, the engine will inherently use the
first non-nullable UNIQUE KEY as the clustered index key (implicit PK), if one
exists. In this case, there's no real harm to performance, and tools like
pt-osc can still be used on the table. For this reason, lint-pk displays a
shorter message in this situation; some DBAs are OK with this situation while
others still prefer an explicit PK for clarity's sake. In any case, no changes
to that behavior in this commit.

However, the situation where a UNIQUE KEY is present, but non-usable as
clustered index key due to allowing NULLs, can be confusing since it's far
from obvious. In this case, tools like pt-osc cannot even be used. This commit
adds additional explanation text to lint-pk to make this more clear.

This commit also updates the vendored version of Go La Tengo to bring in a
semi-related fix for diffing non-InnoDB tables.