MySQL 8: improve robustness of utf8mb3 handling
In MySQL 8.0.24 (April 2021), SHOW CREATE TABLE started displaying the legacy
3-byte "utf8" character set as "utf8mb3" for the table-level DEFAULT CHARSET
clause, but left all other occurrences as "utf8", including in
information_schema.

Skeema has handled this change properly since 44e227f (also April 2021),
however the approach used was not particularly future-proof. In particular,
we expect MySQL 8 will soon switch to using "utf8mb3" in place of "utf8" in
all locations; the MySQL manual actually erroneously states in two places
that this already happened in 8.0.28:

https://dev.mysql.com/doc/refman/8.0/en/charset-unicode-utf8mb3.html
https://dev.mysql.com/doc/refman/8.0/en/charset-unicode-utf8.html

In reality this is a MySQL documentation bug; no change occurred in 8.0.28.
Nonetheless, this change is likely to happen soon in an upcoming release. This
Skeema commit is aims to ensure this change will be handled properly,
refactoring several aspects of MySQL 8 charset/collation handling in the
process.