workspace=docker: log warning if cleanup fails
When using workspace=docker along with either docker-cleanup=stop or
docker-cleanup=destroy, it is possible that the Docker engine could fail to
successfully stop or remove the container. This does not impact Skeema's
operation (since cleanup occurs at program exit), but it is good to inform
the user so that appropriate cleanup action can be performed manually if
desired. This commit logs a warning in this situation.