diff/push: fix MySQL 8 collation edge case verify failure
MySQL 8 exhibits some unusual behavior in SHOW CREATE TABLE regarding
superfluous column-level CHARACTER SET and COLLATE clauses (where
"superfluous" means "equal to the table-level defaults"). This logic was
consistent in MySQL 5.7 and older, and in all versions of MariaDB; however,
MySQL 8 seems to use different and sometimes-inconsistent logic. This caused
false-positives for diff/push --verify failures beginning with Skeema v1.7.1
since it refactored some related logic in a03718c.

Essentially, the --verify option assumed the output of SHOW CREATE TABLE is
a stable "canonical" representation of the table: if you take the output and
execute it as a CREATE TABLE, and then run SHOW CREATE TABLE on it, you should
get back the same output. Unfortunately this is not always the case in MySQL 8
for tables that use a default collation which differs from the default for the
table's chosen default charset.

To fix --verify, this commit changes the logic away from being solely based on
SHOW CREATE TABLE comparisons, and instead it now does a full table diff
(which inherently first compares SHOW CREATE TABLE, and then compares all
table metadata if those do not match). This comparison is able to detect, but
intentionally ignore, differences in superfluous CHARACTER SET and COLLATE
clauses in MySQL 8.

This commit also adds a new integration test method, which failed on the old
--verify logic in MySQL 8 but now passes.

Fixes #184. Thank you to Etsy's database team for the report!