test suite: skip lower_case_table_names=2 test on Percona Server
Integration test suite coverage for lower_case_table_names=2 relies on bind-
mounting a local MacOS path to the data dir for the test container. This works
properly for MySQL and MariaDB images, but encounters a db init failure on
Percona Server images:

2022-06-02T20:18:35.221215Z 0 [ERROR] Can't start server : Bind on unix socket: Invalid argument
2022-06-02T20:18:35.221313Z 0 [ERROR] Do you already have another mysqld server running on socket: /var/lib/mysql/mysql.sock ?
2022-06-02T20:18:35.221347Z 0 [ERROR] Aborting
MySQL init process in progress...
[repeats many times]
MySQL init process in progress...
MySQL init process failed.

Since this test passes (succeeds) on images of all supported MySQL versions,
it would almost certainly succeed on Percona Server as well, if not for the
container db init issue. So this commit skips this test for Percona Server
images for now. More investigation is required; the failure is only
reproducible when the data dir bind mount goes to a testing.TempDir() but not
a shorter all-lowercase dir.