internal refactor: fs and dumper sub-packages
This commit refactors the types and methods of the fs and dumper sub-packages,
including substantial changes to the division of responsibilities between
them.

This work is being done in preparation for the new lexer/parser logic, which
will land in the coming weeks and involve further sub-package adjustments. In
particular, the new lexer/parser is planned to be part of the tengo sub-
package, rather than fs. This will eventually enable all "parse field X out
of SHOW CREATE TABLE to avoid an information_schema inaccuracy in flavor Y"
code paths in tengo to standardize on the singular efficient lexer/parser
implementation.

This commit does not make any functional changes, but does include these
internal refactors:

* fs.Statement no longer has a field pointing back to a SQL file. This was
  essential to enable the upcoming move of fs.Statement to become
  tengo.Statement in the near future.

* fs.SQLFile and fs.TokenizedSQLFile have been combined and substantially
  adjusted. The new fs.SQLFile has methods that are better-aligned with the
  call patterns required by dumper, simplifying corresponding dumper logic.

* Logic involving the fs.LogicalSchema type has been moved to its own file,
  but remains in the fs package

* fs.SQLFiles is now a map, keyed by full file path, rather than a slice

* fs.Dir now tracks NamedSchemaStatements as a field (populated when the dir
  is first parsed) instead of a method

* When dumper needs to change the delimiter around a new or modified
  statement, it now properly inserts separate fs.Statement values, rather
  than hacking in the delimiter commands to the single statement's text