diff/push: support TABLESPACE clauses for system or file-per-table
MySQL 5.7+ supports the ability to put a table in a specific tablespace,
regardless of the server's innodb_file_per_table setting. Such tables will
have a TABLESPACE clause in their SHOW CREATE TABLE; these TABLESPACE clauses
may also be used in ALTER TABLE.

Previously, Skeema was unable to diff tables with such a clause. This is now
supported, with the limitation that the tablespace must be one of the two
built-in options: `innodb_system` or `innodb_file_per_table`.

The ability to use arbitrary user-defined tablespaces ("general tablespaces")
is still not supported yet, as this will require the ability for Skeema to
manage CREATE TABLESPACE statements.