test suites: adjustments for MySQL 8.0.30
Yesterday's release of MySQL 8.0.30 included a change to how the legacy 3-byte
utf8 *collations* are reported by information_schema and SHOW commands.
Previously in MySQL 8.0.29, the utf8 *character set* was changed to utf8mb3,
but the corresponding collation names were not updated. Now in 8.0.30, the
collation names have been updated as well, which is equivalent to the behavior
in MariaDB 10.6+.

This commit adjusts test suite logic to expect the new collation names if the
server is MySQL 8.0.30+, or a variant such as Percona Server 8.0.30+.